Initial composer framework is done
Right now, the Composer cog has two working commands: the command group `comp` itself, and a `create` command. This lets people make teams and compositions. Right now, querying for teams is possible, but not for compositions. I'll add that in a later commit. But for now, we have some real functionality in this cog.