Implement update feature into `Player`
This was a pain to do. Mostly because you can't pass a json string into
rust-postgres even if it *is* valid JSON. Therefore, I had to change the
type in the database from `JSON` to `TEXT`. The reason will be noted
down in the `data/main_db.sql` file, next to the column declaration.
Otherwise, this works pretty much perfectly.
 Changes committed:
	modified:   data/main_db.sql
	modified:   src/utils/db/models.rs