THE SQL FINALLY WORKS AAAAAAAAAAAAAAAAAAAAAAAAAAAA
OK SO
The story goes:

The initial purpose of the async-await push was to use an async lib
called SQLx. *However*, when I tried to use it, it was being retarded
and did not infer the types correctly. Luckily I was smart enough to
commit the async-await changes before I switched the database libraries.
Going back to `mysql`, I was having *constant* panics because the
library couldn't cast whatever type it was getting to the relevant Rust
type. SPECULATION: What proabably happened was that the integer based
indexes I used were pointing to the wrong values.

With regard to making sure the SQL itself was correct, I just refactored
the statements into standalone files and used DataGrip to verify them
(although this has yet to be done for the team module, but the player
one was more pressing).

I'm just glad it's *fucking over.*

*dies*